requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# FINSOC-BOT/bot/extensions/verify.py: 1,2,3
# FINSOC-BOT/bot/main.py: 1,2
discord.py == 1.7.3

# FINSOC-BOT/bot/extensions/verify.py: 12
# FINSOC-BOT/bot/main.py: 3
python_dotenv == 0.18.0
